ps2: Refactor overlay into separate file
It's too difficult to change the overlay code with it stuffed
into `main.c`.
